chore(GIFT-10025): split up gift validation constants
